… Daqifi.Desktop.Test/Exporter/OptimizedExporterValidationTests.cs
… Daqifi.Desktop.Test/Exporter/PerformanceComparisonTests.cs
… Daqifi.Desktop/Exporter/LoggingSessionExporter.cs
… drawing.md
… drawing2.md